+++ extensions/ui/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 2025-11-30
+
+- Implemented streaming support in chat UI
+- Improved stream buffer handling and error reporting
+
 ## 2025-11-25
 
 - Initial release of UI extension